Refactor SSL code
Put the code into the baseprovider so it can be used by OpenStack and
OpenNebula providers easily. May be worth to put it as a separate class.
